+++ gcovr/formats/json/write.py
@@ -98,10 +98,9 @@ def write_summary_report(covdata, output_file: str, options: Options):
     # Data
     keys = sort_coverage(
         covdata,
-        by_branch=options.sort_branches,
-        by_num_uncovered=options.sort_uncovered,
-        by_percent_uncovered=options.sort_percent,
-        reverse=options.sort_reverse,
+        sort_key=options.sort_key,
+        sort_reverse=options.sort_reverse,
+        by_metric="branch" if options.sort_branches else "line",
     )
 
     for key in keys: